pyproject.toml
[tool.poetry]
name = "homework-django-rest-framework"
version = "0.1.0"
description = ""
authors = ["ludmila davletshina <ludaplv@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.30.0"
coverage = "^7.6.8"


[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
